build: make build pass
mssun/objective-git contains build fixes but I have trouble
building the project with said fixes. Therefore, I reverted
to the upstream library and got a successful build. Upstream
is ahead by 20 commits but I don't find any feature patches.
It should be safe to upgrade.

Carthage downloads the binary as upstream made it available.
I'm not sure if this is a licensing problem.